Moved functionality out of main.c
